% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dev.R
\name{calculate_variance}
\alias{calculate_variance}
\title{standard_dev}
\usage{
calculate_variance(ppr)
}
\arguments{
\item{ppr}{An object of class ppr.}
}
\value{
the standard deviaiton of the predicted position of cells.
}
\description{
Calculate the standard deviaiton in the predicted position of cells.
}
