% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{calculate_zscore}
\alias{calculate_zscore}
\title{zscore}
\usage{
calculate_zscore(sce, ppr, switching_genes)
}
\arguments{
\item{sce}{A Single Cell Experiment object,
containing a matrix of your samples binary gene expression,
which has been filtered to only include switching genes,
using PathPinPointR::subset_switching_genes().}

\item{ppr}{An object of class ppr, must be the same sample.}

\item{switching_genes}{Genes which switch through the trajectory,}
}
\value{
the zscore of the raw ppr score.
}
\description{
Calculate the zscore of the raw ppr score.
}
