% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_plot.R
\name{cell_plot}
\alias{cell_plot}
\title{PPR INDIVIDUAL CELL PLOT}
\usage{
cell_plot(
  sample_ppr,
  cell_idx = 1,
  col = "red",
  overlay = FALSE,
  genes_of_interest = NULL,
  switching_genes = NULL,
  true_position = FALSE,
  accuracy_data = NULL
)
}
\arguments{
\item{sample_ppr}{A ppr_OBJECT of the sample you wish to plot}

\item{cell_idx}{The selected cell.}

\item{col}{The colour that you'd like}

\item{overlay}{TRUE if you would like this plot to overlay a previous plot.}

\item{genes_of_interest}{The genes that you would like to plot}

\item{switching_genes}{The data which includes all of the switching genes}

\item{true_position}{do you want to plot the "true position of your cell"}

\item{accuracy_data}{dataset which contains the accuracy data}
}
\value{
plot of the probable position of your sample on your trajectory.
}
\description{
Plots the predicted position of a chosen cell.
}
