% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_position.R
\name{cell_position_box}
\alias{cell_position_box}
\title{cell_position_box}
\usage{
cell_position_box(sample_ppr, col = "gray", points = FALSE)
}
\arguments{
\item{sample_ppr}{A PPR_OBJECT}

\item{col}{A character vector specifying the colour for the sample.}

\item{points}{A logical indicating whether to add points to the boxplot.}
}
\value{
Adds boxplots to the PPR plot, visualising the predicted pseudotime
positions for each cell in a sample.
Points indicate the predicted position of each cell in the sample.
}
\description{
Creates a boxplot overlay showing the predicted pseudotime
positions for each cell in a sample.
}
