% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onAttach}
\alias{.onAttach}
\title{welcome_to_PPR}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The name of the package library.}

\item{pkgname}{The name of the package.}
}
\value{
"Welcome to PathPinpointR!"
}
\description{
prints a welcome message when the package is loaded
}
