% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_example_data}
\alias{get_example_data}
\title{get_example_data}
\usage{
get_example_data()
}
\value{
example data is saved to the current working directory
}
\description{
downloads example data from dropbox
}
