% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_synthetic_data}
\alias{get_synthetic_data}
\title{get_synthetic_data}
\usage{
get_synthetic_data()
}
\value{
a list of three single cell experiment objects;
reference_sce, sample 1 and sample 2.
}
\description{
genetates a synthetic trajectory,
and generates two subsets of the trajectory to be used as example samples.
}
