% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_output_plot.R
\name{gss_output_plot}
\alias{gss_output_plot}
\title{GSS OUTPUT PLOT}
\usage{
gss_output_plot(
  sample.gss,
  col = "red",
  overlay = FALSE,
  label = "sample name",
  genes_of_interest = NULL
)
}
\arguments{
\item{sample.gss}{A GSS_OBJECT of the sample you wish to plot}

\item{col}{The colour that you'd like}

\item{overlay}{set to TRUE if you would like this plot to overlay a previous plot.}
}
\value{
nice plot highlighting the probable position of your sample on your trajectory.
}
\description{
Plots the predicted positon of your sample.
}
