% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_plot_cell_range.R
\name{gss_plot_cell_range}
\alias{gss_plot_cell_range}
\title{gss_plot_cell_range}
\usage{
gss_plot_cell_range(
  sample.gss,
  cell_idxs = c(1:2, 4),
  genes_of_interest = NULL,
  switching_genes = NULL
)
}
\arguments{
\item{sample.gss}{GSS OBJECT}

\item{cell_idxs}{Range or list of cell indexes}

\item{genes_of_interest}{The genes that you would like to plot}

\item{switching_genes}{The data which includes all of the switching genes}
}
\value{
plot to show a large number of cell positions
}
\description{
gss_plot_cell_range
}
