% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_position.R
\name{plot_position}
\alias{plot_position}
\title{PLOT POSITION}
\usage{
plot_position(
  sample_ppr,
  col = "red",
  overlay = FALSE,
  label = "sample name",
  metrics = FALSE,
  genes_of_interest = NULL,
  switching_genes,
  raw_values = FALSE
)
}
\arguments{
\item{sample_ppr}{A PPR_OBJECT of the sample you wish to plot}

\item{col}{The colour that you'd like}

\item{overlay}{TRUE if you would like this plot to overlay a previous plot.}

\item{label}{string that you would like to assign as the label to the line.}

\item{metrics}{logical,
TRUE to include zscore pvalue and SD in the plot.}

\item{genes_of_interest}{character,
The names of any genes that you'd like to include the switching point of.}

\item{switching_genes}{a matrix containing the switching gene information,
as produced by GeneSwitches.}

\item{raw_values}{logical,
TRUE if you want to plot the raw values instead of percentages.}
}
\value{
a plot which highlights the probable position of your sample,
on your trajectory.
}
\description{
Plots the predicted position of your sample.
}
