% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppr_accuracy_test.R
\name{pppr_accuracy_test}
\alias{pppr_accuracy_test}
\title{Title Accuracy Tester}
\usage{
pppr_accuracy_test(reference.gss, reference.gs)
}
\arguments{
\item{reference.gss}{A "GSS_OBJECT" as outputted by create_racing_lines}

\item{reference.gs}{The GS object containing Pseudotime.}
}
\value{
A score for how accurate the results are as a data frame.
}
\description{
Use this with "sample" cells taken from the reference. It will check how close the predicted position of the cells is to the real position.
}
