% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppr_cell_plot.R
\name{pppr_cell_plot}
\alias{pppr_cell_plot}
\title{GSS INDERVIDUAL CELL PLOT}
\usage{
pppr_cell_plot(
  sample.pppr,
  cell_idx = 1,
  col = "red",
  overlay = FALSE,
  genes_of_interest = NULL,
  switching_genes = NULL,
  true_position = FALSE,
  accuracy_data = NULL
)
}
\arguments{
\item{sample.pppr}{A PPPR_OBJECT of the sample you wish to plot}

\item{cell_idx}{The selected cell.}

\item{col}{The colour that you'd like}

\item{overlay}{set to TRUE if you would like this plot to overlay a previous plot.}

\item{genes_of_interest}{The genes that you would like to plot}

\item{switching_genes}{The data which includes all of the switching genes}

\item{true_position}{do you wan to plot the "true position of your cell"}

\item{accuracy_data}{dataset which contains the accuracy data}
}
\value{
nice plot highlighting the probable position of your sample on your trajectory.
}
\description{
Plots the predicted positon of a chosen cell.
}
