% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppr_precision.R
\name{pppr_precision}
\alias{pppr_precision}
\title{pppr_precision}
\usage{
pppr_precision(sample.gs, range = seq(0.02, 0.04, 0.005))
}
\arguments{
\item{sample.gs}{GeneSwitches object that you have already run binarise and GLM on.}

\item{range}{range of r2cutoffs to use}
}
\value{
the precision dataframe.
}
\description{
pppr_precision
}
