% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ppr2seu}
\alias{ppr2seu}
\title{ppr2seu}
\usage{
ppr2seu(ppr, seu, colname = "PPR_pseudotime_idx")
}
\arguments{
\item{ppr}{A ppr object.}

\item{seu}{A Seurat object.}

\item{colname}{The name of the column to add to the Seurat object.}
}
\value{
The Seurat object with the predicted position added to the metadata.
}
\description{
This wrapper function extracts the predicted position of each
cell from a ppr object and adds it to the metdata of a Seurat object.
}
