% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_boxplot.R
\name{ppr_boxplot}
\alias{ppr_boxplot}
\title{ppr_boxplot}
\usage{
ppr_boxplot(samples_ppr, reference_sce, ident)
}
\arguments{
\item{samples_ppr}{A list of PPR objects or a single PPR object.}

\item{reference_sce}{A SingleCellExperiemnt object of the ref trajectory.}

\item{ident}{selected column from reference_sce colData.}
}
\value{
Creates a boxplot showing the predicted pseudotime
positions for each cell in a sample. As well as boxplots of reference cells,
subset by ident.
}
\description{
Creates a boxplot showing the predicted pseudotime
positions for each cell in a sample. As well as boxplots of reference cells,
subset by ident.
}
