% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_output_plot.R
\name{ppr_output_plot}
\alias{ppr_output_plot}
\title{PPR OUTPUT PLOT}
\usage{
ppr_output_plot(
  sample.gss,
  col = "red",
  overlay = FALSE,
  label = "sample name",
  genes_of_interest = NULL,
  switching_genes
)
}
\arguments{
\item{sample.gss}{A GSS_OBJECT of the sample you wish to plot}

\item{col}{The colour that you'd like}

\item{overlay}{set to TRUE if you would like this plot to overlay a previous plot.}

\item{label}{string that you would like to assign as the label to the line.}

\item{genes_of_interest}{The names of any genes that you'd like to include the switching point of.}

\item{switching_genes}{a matrix containing the switching gene information as produced by GeneSwitches.}
}
\value{
nice plot highlighting the probable position of your sample on your trajectory.
}
\description{
Plots the predicted positon of your sample.
}
