% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_predict_position.R
\name{ppr_predict_position}
\alias{ppr_predict_position}
\title{PPPR Predict Position}
\usage{
ppr_predict_position(reduced_binary_counts_matrix, reference.sg)
}
\arguments{
\item{reduced_binary_counts_matrix}{a matrix of your samples binary gene expression.}

\item{reference.sg}{Genes which switch through the trajectory as identified by GeneSwitches.}
}
\value{
A list of matrices: A matrix for each cell where the columns represent progress through a trajectory,
and the rows represent genes, values indicate a likely position of the cell upon the trajectory based that genes bianrized expression.
}
\description{
Produces an estimate for the position on trajectory of each gene in each cell of a sample.
This can be later aggregated to estimate the position of the sample along the trajectory.
}
