% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_racinglines_plot.R, R/ppr_timeline_plot.R
\name{ppr_timeline_plot}
\alias{ppr_timeline_plot}
\title{Identify and Visualise each cells position.}
\usage{
ppr_timeline_plot(
  switching_genes,
  genomic_expression_traces = FALSE,
  reduced_sce = NULL,
  cell_id = 1
)

ppr_timeline_plot(
  switching_genes,
  genomic_expression_traces = FALSE,
  reduced_sce = NULL,
  cell_id = 1
)
}
\arguments{
\item{switching_genes}{A selection of switching genes.}

\item{genomic_expression_traces}{Logical,
Do you want lines which indicate the predicted position of the selected cell.}

\item{reduced_sce}{a SingleCellExperiment object,
that you have already run PathPinPointR::subset_switching_genes() on.}

\item{cell_id}{The index or name of the cell of interest}

\item{reference.sg}{A selection of switching genes which are evenly distributed through pseudo-time.}

\item{lines}{logical, do you want to include lines to indicate the predicted position of a cell}

\item{reference_reduced}{a matrix of your sample's binary gene expression.}
}
\value{
Timeline plot of selected cell

Timeline plot of selected cell
}
\description{
Produces a plot for each cell which helps visualize how PathPinPointR is predicting the cells position.

Produces a plot for a given cell,
this helps visualize predicted position of the selected cell.
}
