% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_vioplot.R
\name{ppr_vioplot}
\alias{ppr_vioplot}
\title{ppr_vioplot}
\usage{
ppr_vioplot(samples_ppr, reference_sce, ident)
}
\arguments{
\item{samples_ppr}{A list of PPR objects or a single PPR object.}

\item{reference_sce}{A SingleCellExperiemnt object of the ref trajectory.}

\item{ident}{selected column from reference_sce colData.}
}
\value{
Creates a violin plot showing the predicted pseudotime
positions for each cell in a sample. As well as violin plots of reference cells,
subset by ident.
}
\description{
Creates a violin plot showing the predicted pseudotime
positions for each cell in a sample. As well as violin plots of reference cells,
subset by ident.
}
