% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\title{precision}
\usage{
precision(sce, n_sg_range = seq(0, 500, 25), plot = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiement object,
that you have already run GeneSwitches::binarize_exp(),
and GeneSwitches::find_switch_logistic_fastglm() on.}

\item{n_sg_range}{range of topnum values to use}

\item{plot}{logical, do you want to plot the precision df.}
}
\value{
a plot of the precision df.
}
\description{
Used to find the optimum topnum value to use,
#' based on using reference as your sample.
}
