% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisionP.r
\name{precisionP}
\alias{precisionP}
\title{precisionP}
\usage{
precisionP(sce, n_sg_range = seq(0, 500, 25), plot = TRUE, cpu = 1)
}
\arguments{
\item{sce}{A SingleCellExperiment object that has already undergone the following preprocessing steps:
\itemize{
\item \code{GeneSwitches::binarize_exp()}
\item \code{GeneSwitches::find_switch_logistic_fastglm()}
}}

\item{n_sg_range}{A numeric vector specifying the range of \code{topnum} values to test. Default is \code{seq(0, 500, 25)}.}

\item{plot}{A logical value indicating whether to plot the precision results. Default is \code{TRUE}. If \code{FALSE}, the function
will return a data frame with the precision values.}

\item{cpu}{An integer specifying the number of CPU cores to use for parallel processing. Default is \code{1}. If \code{cpu} is 0 or exceeds
the number of available cores, it will default to using all but one core.}
}
\value{
If \code{plot = TRUE}, the function returns a plot showing the mean inaccuracy for each value in \code{n_sg_range}.
If \code{plot = FALSE}, the function returns a data frame with columns \code{n_sg} (the number of switching genes) and \code{inaccuracy} (the mean inaccuracy).
}
\description{
This function computes the precision (mean inaccuracy) for different values of topnum (number of switching genes)
by using a reference sample. It runs in parallel to speed up the computation when large datasets are used.
The function returns either a plot of precision (mean inaccuracy) versus the number of switching genes
or a data frame containing the computed precision values for each topnum.
}
