% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_position.R
\name{predict_position}
\alias{predict_position}
\title{predict_position}
\usage{
predict_position(sample_sce, switching_genes)
}
\arguments{
\item{sample_sce}{A Single Cell Experiment object,
containing a matrix of your samples binary gene expression,
which has been filtered to only include switching genes,
using PathPinPointR::subset_switching_genes().}

\item{switching_genes}{Genes which switch through the trajectory,
as identified by GeneSwitches.}
}
\value{
A list of matrices:
A matrix for each cell:
the columns represent progress through a trajectory,
the rows represent genes,
values indicate a likely position of the cell upon the trajectory,
based that genes bianrized expression.
}
\description{
Produces an estimate for the position on trajectory of each cell of a sample.
This is later aggregated to estimate samples position along the trajectory.
}
