% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppr_utils.R
\name{print.PPPR_OBJECT}
\alias{print.PPPR_OBJECT}
\title{Print PPPR Object Summary}
\usage{
\method{print}{PPPR_OBJECT}(x)
}
\arguments{
\item{x}{An object of class 'PPPR_OBJECT' to be summarized.}
}
\value{
A summary of the contents of the 'PPPR_OBJECT'.
}
\description{
Make calling the PPPR_OBJECT nicer by printing a summary of its contents.
}
