% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.PPR_OBJECT}
\alias{print.PPR_OBJECT}
\title{Print method for PPR Object.}
\usage{
\method{print}{PPR_OBJECT}(x, ...)
}
\arguments{
\item{x}{An object of class 'PPR_OBJECT' to be summarized.}

\item{...}{Additional arguments passed to print (unused).}
}
\value{
Invisibly returns the input object. Called for its side effect of printing a summary.
}
\description{
Make calling the PPR_OBJECT nicer by printing a summary of its contents.
This function can be used to quickly inspect the structure of your results.
}
