% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_bias.R
\name{pt_bias}
\alias{pt_bias}
\title{pt_bias}
\usage{
pt_bias(ppr, reference_sce)
}
\arguments{
\item{ppr}{A "PPR_OBJECT" which originates from the reference,
Or a subset of the reference.}

\item{reference_sce}{A SingleCellExperiment object.
This object should contain:
\itemize{
\item \code{Pseudotime}: A numeric vector representing the progression of cells along a trajectory.
\item \code{Binarized}: A matrix or data frame indicating the binary expression status of genes.
}}
}
\value{
A ggplot object showing a scatter plot of predicted vs. true pseudotime indices,
with a reference line indicating perfect prediction.
These elements should originate from the reference data.

returns a ggplot object showing the bias of PPR predictions across pseudo-time.
}
\description{
This function plots the accuracy of PPR predictions across pseudo-time.
}
