% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_counts_matrix.R
\name{reduce_counts_matrix}
\alias{reduce_counts_matrix}
\title{reduce_counts_matrix}
\usage{
reduce_counts_matrix(sample_sce, switching_genes)
}
\arguments{
\item{sample_sce}{A Single Cell Experiment object,
containing a matrix of your samples gene expression.}

\item{switching_genes}{Genes which switch through the trajectory,
as identified by GeneSwitches.}
}
\value{
sample_sce A Single Cell Experiment object,
containing a reduced matrix of your samples gene expression,
only genes included in switching_gene are included.
}
\description{
Reduces the counts matrix to only include genes included in switching_genes.
}
