% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_position.R
\name{reference_idents}
\alias{reference_idents}
\title{reference_idents}
\usage{
reference_idents(reference_sce, ident)
}
\arguments{
\item{reference_sce}{A SingleCellExperiemnt object of the ref trajectory}

\item{ident}{selected column from reference_sce colData.}
}
\value{
overlay for the plot_position function with reference idents.
}
\description{
Overlay to ppr plot visualising the pseudotime of reference cells,
subset by ident.
}
