% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_plot.R
\name{sample_prediction}
\alias{sample_prediction}
\title{sample_prediction}
\usage{
sample_prediction(sample_ppr, col = "red", label = "sample name")
}
\arguments{
\item{sample_ppr}{A PPR_OBJECT of the sample you wish to plot}

\item{col}{The colour that you'd like}

\item{label}{string that you would like to assign as the label to the line.}
}
\value{
an overlay for the ppr plot,
which highlights the probable position of your sample in pseudotime.
}
\description{
Plots the predicted position of your sample.
}
