% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_test.R
\name{score_gss_accuracy}
\alias{score_gss_accuracy}
\title{Title Accuracy Tester}
\usage{
score_gss_accuracy(gss_obj, reference.sce)
}
\arguments{
\item{gss_obj}{A "GSS_OBJECT" as outputted by create_racing_lines}

\item{reference.sce}{The SCE object which has had slingshot run on it already.}
}
\value{
A score for how accurate the results are as a data frame.
}
\description{
Use this with "sample" cells taken from the reference. It will check how close the predicted position of the cells is to the real position.
}
