% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_switching_genes.R
\name{subset_switching_genes}
\alias{subset_switching_genes}
\title{Subset Gene Expression for Switching Genes}
\usage{
subset_switching_genes(sample_sce, switching_genes)
}
\arguments{
\item{sample_sce}{a Single Cell Experiment object,
which contains a binarized expression matrix,
this is your sample.}

\item{switching_genes}{Genes identified by GeneSwitches as "switching",
produced using your reference.}
}
\value{
a binary expression matrix subseted to only include switching genes
}
\description{
Reduce a binary expression matrix to only the switching genes.
}
