% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_plot.R
\name{switching_times}
\alias{switching_times}
\title{switching_times}
\usage{
switching_times(genes, switching_genes)
}
\arguments{
\item{genes}{A character vector of gene names you wish to plot.}

\item{switching_genes}{as output from GeneSwitches::find_switch_logistic_fastglm()}
}
\value{
Adds dashed lines and gene names to the PPR plot, indicating
pseudotime switching points of interest.
}
\description{
Adds gene switching points to a PPR plot.
The function overlays dashed lines at the pseudotime points where genes
switch on or off, along with labels for each gene.
}
