% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\name{timeline_plot}
\alias{timeline_plot}
\title{Identify and Visualise each cells position.}
\usage{
timeline_plot(
  switching_genes,
  genomic_expression_traces = FALSE,
  reduced_sce = NULL,
  cell_id = 1
)
}
\arguments{
\item{switching_genes}{A selection of switching genes.}

\item{genomic_expression_traces}{Logical,
Do you want lines which indicate the predicted position of the selected cell.}

\item{reduced_sce}{a SingleCellExperiment object,
that you have already run PathPinpointR::subset_switching_genes() on.}

\item{cell_id}{The index or name of the cell of interest}
}
\value{
Timeline plot of selected cell
}
\description{
Produces a plot for a given cell,
this helps visualize predicted position of the selected cell.
}
