% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_synthetic_data}
\alias{get_synthetic_data}
\title{get_synthetic_data}
\usage{
get_synthetic_data()
}
\value{
A list of three SingleCellExperiment objects: one reference and two samples.
}
\description{
Generates a synthetic trajectory and two subsets of the trajectory
to be used as example samples.
}
