% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppr_timeline_plot.R
\name{pppr_timeline_plot}
\alias{pppr_timeline_plot}
\title{Identify and Visualise each cells position.}
\usage{
pppr_timeline_plot(reference.sg, reference_reduced, cell_idx = 1, lines = TRUE)
}
\arguments{
\item{reference.sg}{A selection of switching genes which are evenly distributed through pseudo-time.}

\item{reference_reduced}{a matrix of your samples binary gene expression.}

\item{cell_idx}{The index (should get changed to name) of the cell of interest}

\item{lines}{Logical, Do you want to plot the lines which indicate the predicted position of the selected cell.}
}
\value{
Timeline plot of selected cell
}
\description{
Produces a plot for each cell which helps visualize how GSS is predicting the cells position.
}
