% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppr_timeline_plot.R
\name{pppr_timeline_plot}
\alias{pppr_timeline_plot}
\title{Identify and Visualise each cells position.}
\usage{
pppr_timeline_plot(
  reference.sg,
  lines = FALSE,
  reduced_binary_counts_matrix = NULL,
  cell_idx = 1
)
}
\arguments{
\item{reference.sg}{A selection of switching genes.}

\item{lines}{Logical, Do you want to plot the lines which indicate the predicted position of the selected cell.}

\item{reduced_binary_counts_matrix}{a matrix of your samples binary gene expression.}

\item{cell_idx}{The index (should get changed to name) of the cell of interest}
}
\value{
Timeline plot of selected cell
}
\description{
Produces a plot for each cell which helps visualize how GSS is predicting the cells position.
}
