% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_accuracy_test.R
\name{ppr_accuracy_test}
\alias{ppr_accuracy_test}
\title{accuracy_test}
\usage{
ppr_accuracy_test(reference_ppr, reference_sce, plot = TRUE)
}
\arguments{
\item{reference_ppr}{A "PPR_OBJECT" as outputted by create_racing_lines}

\item{reference_sce}{A SingleCellExperiment object,
containing Pseudotime, binarized and found switching genes.}

\item{plot}{Logical value. Default is FALSE.
If TRUE, a histogram of inaccuracy will be plotted.}
}
\value{
If plot = FALSE, returns a data frame containing accuracy results.
If plot = TRUE, returns a the accuracy data frame.
}
\description{
Use this with "sample" cells taken from the reference.
Checks how close the predicted position of the cells is to the true position.
}
