% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\title{precision}
\usage{
precision(sce, r2_cutoff_range = seq(0, 0.5, 0.1), plot = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiement object,
that you have already run GeneSwitches::binarize_exp(),
and GeneSwitches::find_switch_logistic_fastglm() on.}

\item{r2_cutoff_range}{range of r2cutoffs to use}

\item{plot}{logical, do you want to plot the precision df.}
}
\value{
a plot of the precision df.
}
\description{
Used to find the optimum r2cutoff to use,
#' based on using reference as your sample.
}
