% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPR_timeline_plot.R, R/plot_racinglines.R
\name{PPR_timeline_plot}
\alias{PPR_timeline_plot}
\title{Identify and Visualise each cells position.}
\usage{
PPR_timeline_plot(
  reference.sg,
  lines = TRUE,
  reference_reduced = NULL,
  cell_idx = 1,
  cell_name = NULL
)

PPR_timeline_plot(
  reference.sg,
  lines = TRUE,
  reference_reduced = NULL,
  cell_idx = 1,
  cell_name = NULL
)
}
\arguments{
\item{reference.sg}{A selection of switching genes which are evenly distributed through pseudo-time.}

\item{reference_reduced}{a matrix of your samples binary gene expression.}

\item{cell_idx}{The index (should get changed to name) of the cell of interest}

\item{full_time_IDX}{Do you want the scale to go from Min to Max or from 0-100.}
}
\value{
Timeline plot of selected cell

Timeline plot of selected cell
}
\description{
Produces a plot for each cell which helps visualize how GSS is predicting the cells position.

Produces a plot for each cell which helps visualize how GSS is predicting the cells position.
}
