% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.PPR_OBJECT}
\alias{print.PPR_OBJECT}
\title{Print method for PPR Object.}
\usage{
print.PPR_OBJECT(x)
}
\arguments{
\item{x}{An object of class 'PPR_OBJECT' to be summarized.}
}
\value{
A summary of the contents of the object.
}
\description{
Make calling the PPR_OBJECT nicer by printing a summary of its contents.
}
