% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_utils.R
\name{which_mid_max}
\alias{which_mid_max}
\title{Select the middle index among multiple occurrences of the maximum value}
\usage{
which_mid_max(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
The middle index among multiple occurrences of the maximum value.
}
\description{
Select the middle index among multiple occurrences of the maximum value
}
