% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_mid_max}
\alias{which_mid_max}
\title{Select the mid index among multiple occurrences of the max value.}
\usage{
which_mid_max(n)
}
\arguments{
\item{n}{Numeric vector.}
}
\value{
The middle index among multiple occurrences of the maximum value.
}
\description{
This helper function is used within PathPinpointR to resolve ties when
multiple values are equal to the maximum.
}
